# core/vm: faster code analysis