--- a/core/vm/analysis.go
+++ b/core/vm/analysis.go
@@ -21,12 +21,48 @@ package vm
 // it's data (i.e. argument of PUSHxx).
 type bitvec []byte
 
-func (bits *bitvec) set(pos uint64) {
-	(*bits)[pos/8] |= 0x80 >> (pos % 8)
+var lookup = [8]byte{
+	0x80, 0x40, 0x20, 0x10, 0x8, 0x4, 0x2, 0x1,
 }
+
+func (bits *bitvec) set(pos uint64) {
+	//(*bits)[pos/8] |= 0x80 >> (pos % 8)
+	(*bits)[pos/8] |= lookup[pos%8]
+}
+
+func (bits *bitvec) set2(pos uint64) {
+	(*bits)[pos/8+1] |= 0b1100_0000 << (8 - pos%8)
+	(*bits)[pos/8] |= 0b1100_0000 >> (pos % 8)
+}
+
+func (bits *bitvec) set3(pos uint64) {
+	(*bits)[pos/8+1] |= 0b1110_0000 << (8 - pos%8)
+	(*bits)[pos/8] |= 0b1110_0000 >> (pos % 8)
+}
+
+func (bits *bitvec) set4(pos uint64) {
+	(*bits)[pos/8+1] |= 0b1111_0000 << (8 - pos%8)
+	(*bits)[pos/8] |= 0b1111_0000 >> (pos % 8)
+}
+
+func (bits *bitvec) set5(pos uint64) {
+	(*bits)[pos/8+1] |= 0b1111_1000 << (8 - pos%8)
+	(*bits)[pos/8] |= 0b1111_1000 >> (pos % 8)
+}
+
+func (bits *bitvec) set6(pos uint64) {
+	(*bits)[pos/8+1] |= 0b1111_1100 << (8 - pos%8)
+	(*bits)[pos/8] |= 0b1111_1100 >> (pos % 8)
+}
+
+func (bits *bitvec) set7(pos uint64) {
+	(*bits)[pos/8+1] |= 0b1111_1110 << (8 - pos%8)
+	(*bits)[pos/8] |= 0b1111_1110 >> (pos % 8)
+}
+
 func (bits *bitvec) set8(pos uint64) {
-	(*bits)[pos/8] |= 0xFF >> (pos % 8)
 	(*bits)[pos/8+1] |= ^(0xFF >> (pos % 8))
+	(*bits)[pos/8] |= 0xFF >> (pos % 8)
 }
 
 // codeSegment checks if the position is in a code segment.
@@ -49,21 +85,33 @@ func codeBitmap(code []byte) bitvec {
 func codeBitmapInternal(code, bits bitvec) bitvec {
 	for pc := uint64(0); pc < uint64(len(code)); {
 		op := OpCode(code[pc])
-
-		if op >= PUSH1 && op <= PUSH32 {
-			numbits := op - PUSH1 + 1
-			pc++
-			for ; numbits >= 8; numbits -= 8 {
-				bits.set8(pc) // 8
-				pc += 8
-			}
-			for ; numbits > 0; numbits-- {
-				bits.set(pc)
-				pc++
-			}
-		} else {
-			pc++
+		pc++
+		if op < PUSH1 || op > PUSH32 {
+			continue
 		}
+		numbits := op - PUSH1 + 1
+		for ; numbits >= 8; numbits -= 8 {
+			bits.set8(pc) // 8
+			pc += 8
+		}
+		// numbits now max 7
+		switch numbits {
+		case 1:
+			bits.set(pc)
+		case 2:
+			bits.set2(pc)
+		case 3:
+			bits.set3(pc)
+		case 4:
+			bits.set4(pc)
+		case 5:
+			bits.set5(pc)
+		case 6:
+			bits.set6(pc)
+		case 7:
+			bits.set7(pc)
+		}
+		pc += uint64(numbits)
 	}
 	return bits
 }