This makes the analysis for code/data faster, and more evenly spread. Previously, the worst-performing case was PUSH7.
`PUSH1` started at ~40ns, and going upwards towards `50ms` at `PUSH7`. 
This PR doesn't improve the PUSH1 situation, but does markedly improve the (previous) worst-case quite a lot, and also the normal-case (for code which is not push). 
 
```
name                           old time/op  new time/op  delta
JumpdestOpAnalysis/PUSH1-6     39.7ns ± 3%  40.2ns ± 0%     ~     (p=0.151 n=5+5)
JumpdestOpAnalysis/PUSH2-6     39.4ns ± 0%  38.6ns ± 3%     ~     (p=0.190 n=4+5)
JumpdestOpAnalysis/PUSH3-6     41.4ns ± 1%  31.5ns ± 3%  -23.83%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH4-6     44.5ns ± 6%  23.7ns ± 5%  -46.72%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH5-6     47.7ns ± 5%  22.0ns ± 1%  -53.95%  (p=0.016 n=5+4)
JumpdestOpAnalysis/PUSH6-6     47.2ns ± 1%  20.3ns ± 1%  -57.03%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH7-6     48.7ns ± 2%  18.1ns ± 5%  -62.91%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH8-6     12.0ns ± 2%  19.4ns ±12%  +60.79%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH9-6     15.2ns ± 0%  15.4ns ± 4%     ~     (p=0.460 n=4+5)
JumpdestOpAnalysis/PUSH10-6    18.8ns ± 1%  19.0ns ± 1%     ~     (p=0.095 n=5+5)
JumpdestOpAnalysis/PUSH11-6    22.2ns ± 9%  18.0ns ± 4%  -18.52%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH12-6    24.1ns ± 5%  16.0ns ± 6%  -33.56%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH13-6    25.9ns ± 4%  15.3ns ± 4%  -40.92%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH14-6    27.2ns ± 0%  15.0ns ± 2%  -44.80%  (p=0.016 n=4+5)
JumpdestOpAnalysis/PUSH15-6    29.1ns ± 0%  13.6ns ± 1%  -53.15%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH16-6    11.0ns ± 2%  13.7ns ± 5%  +24.60%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH17-6    12.9ns ± 1%  13.0ns ± 2%     ~     (p=0.905 n=4+5)
JumpdestOpAnalysis/PUSH18-6    15.1ns ± 0%  15.3ns ± 1%   +0.75%  (p=0.016 n=4+5)
JumpdestOpAnalysis/PUSH19-6    16.8ns ± 1%  15.0ns ± 1%  -10.72%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH20-6    19.8ns ±15%  14.1ns ± 6%  -28.64%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH21-6    21.5ns ±20%  13.3ns ± 0%  -38.33%  (p=0.016 n=5+4)
JumpdestOpAnalysis/PUSH22-6    21.2ns ± 1%  13.4ns ± 5%  -36.65%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH23-6    23.0ns ± 2%  12.8ns ± 3%  -44.43%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH24-6    10.8ns ± 3%  12.9ns ± 3%  +19.24%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH25-6    12.1ns ± 1%  12.6ns ± 5%   +4.83%  (p=0.016 n=4+5)
JumpdestOpAnalysis/PUSH26-6    13.4ns ± 0%  14.2ns ± 1%   +5.45%  (p=0.016 n=4+5)
JumpdestOpAnalysis/PUSH27-6    15.1ns ± 0%  13.7ns ± 0%   -9.64%  (p=0.016 n=5+4)
JumpdestOpAnalysis/PUSH28-6    16.1ns ± 2%  13.2ns ± 6%  -18.02%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH29-6    17.5ns ± 6%  12.9ns ± 2%  -26.38%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH30-6    18.4ns ± 0%  13.0ns ± 7%  -29.39%  (p=0.016 n=4+5)
JumpdestOpAnalysis/PUSH31-6    19.7ns ± 1%  12.3ns ± 1%  -37.36%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH32-6    10.6ns ± 2%  12.5ns ± 5%  +17.95%  (p=0.008 n=5+5)
JumpdestOpAnalysis/JUMPDEST-6  38.9ns ± 5%  20.1ns ± 1%  -48.29%  (p=0.008 n=5+5)
JumpdestOpAnalysis/STOP-6      38.8ns ± 4%  20.3ns ± 5%  -47.68%  (p=0.008 n=5+5)
```