@holiman 2021-08-11
Last commit improves the worst-performing case somewhat. 
Difference without and with the last commit: 
```
[user@work vm]$ benchstat final2.txt final3.txt 
name                           old time/op  new time/op  delta
JumpdestOpAnalysis/PUSH1-6     40.2ns ± 0%  31.0ns ± 1%  -22.95%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH2-6     38.6ns ± 3%  38.8ns ± 8%     ~     (p=0.333 n=5+5)
JumpdestOpAnalysis/PUSH3-6     31.5ns ± 3%  28.6ns ± 6%   -9.38%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH4-6     23.7ns ± 5%  25.1ns ± 7%     ~     (p=0.222 n=5+5)
JumpdestOpAnalysis/PUSH5-6     22.0ns ± 1%  21.4ns ± 4%     ~     (p=0.190 n=4+5)
JumpdestOpAnalysis/PUSH6-6     20.3ns ± 1%  19.0ns ± 3%   -6.31%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH7-6     18.1ns ± 5%  17.7ns ± 2%     ~     (p=0.310 n=5+5)
JumpdestOpAnalysis/PUSH8-6     19.4ns ±12%  20.0ns ± 0%     ~     (p=0.730 n=5+4)
JumpdestOpAnalysis/PUSH9-6     15.4ns ± 4%  16.2ns ± 7%     ~     (p=0.056 n=5+5)
JumpdestOpAnalysis/PUSH10-6    19.0ns ± 1%  19.6ns ± 0%   +3.18%  (p=0.016 n=5+4)
JumpdestOpAnalysis/PUSH11-6    18.0ns ± 4%  19.7ns ±11%   +9.19%  (p=0.040 n=5+5)
JumpdestOpAnalysis/PUSH12-6    16.0ns ± 6%  17.0ns ± 5%     ~     (p=0.056 n=5+5)
JumpdestOpAnalysis/PUSH13-6    15.3ns ± 4%  15.6ns ± 5%     ~     (p=0.222 n=5+5)
JumpdestOpAnalysis/PUSH14-6    15.0ns ± 2%  15.7ns ± 6%     ~     (p=0.056 n=5+5)
JumpdestOpAnalysis/PUSH15-6    13.6ns ± 1%  14.4ns ± 2%   +5.48%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH16-6    13.7ns ± 5%  14.1ns ± 0%     ~     (p=0.190 n=5+4)
JumpdestOpAnalysis/PUSH17-6    13.0ns ± 2%  13.9ns ± 4%   +7.57%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH18-6    15.3ns ± 1%  15.4ns ± 5%     ~     (p=0.984 n=5+5)
JumpdestOpAnalysis/PUSH19-6    15.0ns ± 1%  15.2ns ± 0%   +1.41%  (p=0.016 n=5+4)
JumpdestOpAnalysis/PUSH20-6    14.1ns ± 6%  14.5ns ± 2%     ~     (p=0.151 n=5+5)
JumpdestOpAnalysis/PUSH21-6    13.3ns ± 0%  13.9ns ± 6%   +4.64%  (p=0.016 n=4+5)
JumpdestOpAnalysis/PUSH22-6    13.4ns ± 5%  14.1ns ± 7%   +5.09%  (p=0.048 n=5+5)
JumpdestOpAnalysis/PUSH23-6    12.8ns ± 3%  13.4ns ±10%     ~     (p=0.056 n=5+5)
JumpdestOpAnalysis/PUSH24-6    12.9ns ± 3%  12.5ns ± 1%   -2.43%  (p=0.032 n=5+5)
JumpdestOpAnalysis/PUSH25-6    12.6ns ± 5%  12.5ns ± 4%     ~     (p=0.452 n=5+5)
JumpdestOpAnalysis/PUSH26-6    14.2ns ± 1%  14.1ns ± 2%     ~     (p=0.167 n=5+5)
JumpdestOpAnalysis/PUSH27-6    13.7ns ± 0%  13.7ns ± 1%     ~     (p=0.190 n=4+5)
JumpdestOpAnalysis/PUSH28-6    13.2ns ± 6%  13.2ns ± 6%     ~     (p=0.968 n=5+5)
JumpdestOpAnalysis/PUSH29-6    12.9ns ± 2%  12.6ns ± 3%     ~     (p=0.095 n=5+5)
JumpdestOpAnalysis/PUSH30-6    13.0ns ± 7%  12.6ns ± 0%   -3.15%  (p=0.016 n=5+4)
JumpdestOpAnalysis/PUSH31-6    12.3ns ± 1%  12.3ns ± 6%     ~     (p=0.135 n=5+5)
JumpdestOpAnalysis/PUSH32-6    12.5ns ± 5%  12.3ns ± 4%     ~     (p=0.421 n=5+5)
JumpdestOpAnalysis/JUMPDEST-6  20.1ns ± 1%  20.0ns ± 2%     ~     (p=0.222 n=5+5)
JumpdestOpAnalysis/STOP-6      20.3ns ± 5%  19.8ns ± 1%   -2.35%  (p=0.008 n=5+5)
```
And final performance diff against master: 
```
[user@work vm]$ benchstat before.txt final3.txt 
name                           old time/op  new time/op  delta
JumpdestOpAnalysis/PUSH1-6     39.7ns ± 3%  31.0ns ± 1%  -22.01%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH2-6     39.4ns ± 0%  38.8ns ± 8%     ~     (p=0.190 n=4+5)
JumpdestOpAnalysis/PUSH3-6     41.4ns ± 1%  28.6ns ± 6%  -30.97%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH4-6     44.5ns ± 6%  25.1ns ± 7%  -43.65%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH5-6     47.7ns ± 5%  21.4ns ± 4%  -55.22%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH6-6     47.2ns ± 1%  19.0ns ± 3%  -59.74%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH7-6     48.7ns ± 2%  17.7ns ± 2%  -63.62%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH8-6     12.0ns ± 2%  20.0ns ± 0%  +65.75%  (p=0.016 n=5+4)
JumpdestOpAnalysis/PUSH9-6     15.2ns ± 0%  16.2ns ± 7%   +6.39%  (p=0.016 n=4+5)
JumpdestOpAnalysis/PUSH10-6    18.8ns ± 1%  19.6ns ± 0%   +4.40%  (p=0.016 n=5+4)
JumpdestOpAnalysis/PUSH11-6    22.2ns ± 9%  19.7ns ±11%     ~     (p=0.056 n=5+5)
JumpdestOpAnalysis/PUSH12-6    24.1ns ± 5%  17.0ns ± 5%  -29.65%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH13-6    25.9ns ± 4%  15.6ns ± 5%  -39.79%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH14-6    27.2ns ± 0%  15.7ns ± 6%  -42.52%  (p=0.016 n=4+5)
JumpdestOpAnalysis/PUSH15-6    29.1ns ± 0%  14.4ns ± 2%  -50.59%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH16-6    11.0ns ± 2%  14.1ns ± 0%  +28.85%  (p=0.016 n=5+4)
JumpdestOpAnalysis/PUSH17-6    12.9ns ± 1%  13.9ns ± 4%   +7.74%  (p=0.016 n=4+5)
JumpdestOpAnalysis/PUSH18-6    15.1ns ± 0%  15.4ns ± 5%   +1.83%  (p=0.032 n=4+5)
JumpdestOpAnalysis/PUSH19-6    16.8ns ± 1%  15.2ns ± 0%   -9.46%  (p=0.016 n=5+4)
JumpdestOpAnalysis/PUSH20-6    19.8ns ±15%  14.5ns ± 2%  -26.85%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH21-6    21.5ns ±20%  13.9ns ± 6%  -35.47%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH22-6    21.2ns ± 1%  14.1ns ± 7%  -33.42%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH23-6    23.0ns ± 2%  13.4ns ±10%  -41.64%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH24-6    10.8ns ± 3%  12.5ns ± 1%  +16.34%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH25-6    12.1ns ± 1%  12.5ns ± 4%   +3.92%  (p=0.016 n=4+5)
JumpdestOpAnalysis/PUSH26-6    13.4ns ± 0%  14.1ns ± 2%   +4.86%  (p=0.016 n=4+5)
JumpdestOpAnalysis/PUSH27-6    15.1ns ± 0%  13.7ns ± 1%   -9.36%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH28-6    16.1ns ± 2%  13.2ns ± 6%  -17.94%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH29-6    17.5ns ± 6%  12.6ns ± 3%  -28.04%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH30-6    18.4ns ± 0%  12.6ns ± 0%  -31.61%  (p=0.029 n=4+4)
JumpdestOpAnalysis/PUSH31-6    19.7ns ± 1%  12.3ns ± 6%  -37.41%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH32-6    10.6ns ± 2%  12.3ns ± 4%  +16.30%  (p=0.008 n=5+5)
JumpdestOpAnalysis/JUMPDEST-6  38.9ns ± 5%  20.0ns ± 2%  -48.64%  (p=0.008 n=5+5)
JumpdestOpAnalysis/STOP-6      38.8ns ± 4%  19.8ns ± 1%  -48.91%  (p=0.008 n=5+5)