@holiman 2021-08-14
Further incremental improvements: 
```
[user@work vm]$ benchstat final3.txt final4.txt 
name                           old time/op  new time/op  delta
JumpdestOpAnalysis/PUSH1-6     31.0ns ± 1%  30.2ns ± 1%   -2.42%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH2-6     38.8ns ± 8%  29.4ns ± 0%  -24.20%  (p=0.016 n=5+4)
JumpdestOpAnalysis/PUSH3-6     28.6ns ± 6%  24.2ns ± 4%  -15.21%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH4-6     25.1ns ± 7%  20.3ns ± 6%  -19.04%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH5-6     21.4ns ± 4%  19.9ns ± 4%   -6.92%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH6-6     19.0ns ± 3%  17.0ns ± 4%  -10.71%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH7-6     17.7ns ± 2%  15.3ns ± 3%  -13.43%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH8-6     20.0ns ± 0%  15.6ns ± 1%  -22.02%  (p=0.016 n=4+5)
JumpdestOpAnalysis/PUSH9-6     16.2ns ± 7%  14.8ns ± 1%   -8.65%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH10-6    19.6ns ± 0%  15.6ns ± 0%  -20.62%  (p=0.016 n=4+5)
JumpdestOpAnalysis/PUSH11-6    19.7ns ±11%  15.2ns ± 0%  -23.05%  (p=0.016 n=5+4)
JumpdestOpAnalysis/PUSH12-6    17.0ns ± 5%  14.2ns ± 4%  -16.60%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH13-6    15.6ns ± 5%  13.7ns ± 6%  -11.84%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH14-6    15.7ns ± 6%  12.7ns ± 2%  -18.89%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH15-6    14.4ns ± 2%  12.2ns ± 0%  -15.42%  (p=0.016 n=5+4)
JumpdestOpAnalysis/PUSH16-6    14.1ns ± 0%  12.5ns ± 1%  -11.47%  (p=0.016 n=4+5)
JumpdestOpAnalysis/PUSH17-6    13.9ns ± 4%  12.6ns ± 0%   -9.34%  (p=0.016 n=5+4)
JumpdestOpAnalysis/PUSH18-6    15.4ns ± 5%  13.3ns ± 0%  -13.68%  (p=0.016 n=5+4)
JumpdestOpAnalysis/PUSH19-6    15.2ns ± 0%  13.4ns ± 2%  -11.60%  (p=0.016 n=4+5)
JumpdestOpAnalysis/PUSH20-6    14.5ns ± 2%  12.5ns ± 7%  -13.75%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH21-6    13.9ns ± 6%  12.4ns ± 5%  -10.75%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH22-6    14.1ns ± 7%  12.8ns ±13%     ~     (p=0.151 n=5+5)
JumpdestOpAnalysis/PUSH23-6    13.4ns ±10%  12.0ns ± 3%  -10.66%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH24-6    12.5ns ± 1%  12.0ns ± 2%   -4.61%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH25-6    12.5ns ± 4%  12.2ns ± 4%     ~     (p=0.095 n=5+5)
JumpdestOpAnalysis/PUSH26-6    14.1ns ± 2%  12.2ns ± 1%  -13.07%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH27-6    13.7ns ± 1%  12.2ns ± 0%  -11.33%  (p=0.016 n=5+4)
JumpdestOpAnalysis/PUSH28-6    13.2ns ± 6%  12.1ns ± 4%   -8.28%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH29-6    12.6ns ± 3%  11.9ns ± 0%   -5.71%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH30-6    12.6ns ± 0%  11.6ns ± 5%   -7.77%  (p=0.016 n=4+5)
JumpdestOpAnalysis/PUSH31-6    12.3ns ± 6%  11.2ns ± 1%   -9.52%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH32-6    12.3ns ± 4%  11.5ns ± 0%   -6.94%  (p=0.016 n=5+4)
JumpdestOpAnalysis/JUMPDEST-6  20.0ns ± 2%  19.6ns ± 0%   -1.67%  (p=0.016 n=5+4)
JumpdestOpAnalysis/STOP-6      19.8ns ± 1%  19.6ns ± 0%   -1.18%  (p=0.016 n=5+4)
```
Total improvements: 
```
name                           old time/op  new time/op  delta
JumpdestOpAnalysis/PUSH1-6     39.7ns ± 3%  30.2ns ± 1%  -23.90%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH2-6     39.4ns ± 0%  29.4ns ± 0%  -25.43%  (p=0.029 n=4+4)
JumpdestOpAnalysis/PUSH3-6     41.4ns ± 1%  24.2ns ± 4%  -41.47%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH4-6     44.5ns ± 6%  20.3ns ± 6%  -54.38%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH5-6     47.7ns ± 5%  19.9ns ± 4%  -58.32%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH6-6     47.2ns ± 1%  17.0ns ± 4%  -64.05%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH7-6     48.7ns ± 2%  15.3ns ± 3%  -68.51%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH8-6     12.0ns ± 2%  15.6ns ± 1%  +29.25%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH9-6     15.2ns ± 0%  14.8ns ± 1%   -2.81%  (p=0.016 n=4+5)
JumpdestOpAnalysis/PUSH10-6    18.8ns ± 1%  15.6ns ± 0%  -17.13%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH11-6    22.2ns ± 9%  15.2ns ± 0%  -31.54%  (p=0.016 n=5+4)
JumpdestOpAnalysis/PUSH12-6    24.1ns ± 5%  14.2ns ± 4%  -41.33%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH13-6    25.9ns ± 4%  13.7ns ± 6%  -46.92%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH14-6    27.2ns ± 0%  12.7ns ± 2%  -53.37%  (p=0.016 n=4+5)
JumpdestOpAnalysis/PUSH15-6    29.1ns ± 0%  12.2ns ± 0%  -58.20%  (p=0.016 n=5+4)
JumpdestOpAnalysis/PUSH16-6    11.0ns ± 2%  12.5ns ± 1%  +14.08%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH17-6    12.9ns ± 1%  12.6ns ± 0%   -2.32%  (p=0.029 n=4+4)
JumpdestOpAnalysis/PUSH18-6    15.1ns ± 0%  13.3ns ± 0%  -12.10%  (p=0.029 n=4+4)
JumpdestOpAnalysis/PUSH19-6    16.8ns ± 1%  13.4ns ± 2%  -19.96%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH20-6    19.8ns ±15%  12.5ns ± 7%  -36.91%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH21-6    21.5ns ±20%  12.4ns ± 5%  -42.41%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH22-6    21.2ns ± 1%  12.8ns ±13%  -39.48%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH23-6    23.0ns ± 2%  12.0ns ± 3%  -47.86%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH24-6    10.8ns ± 3%  12.0ns ± 2%  +10.98%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH25-6    12.1ns ± 1%  12.2ns ± 4%     ~     (p=0.619 n=4+5)
JumpdestOpAnalysis/PUSH26-6    13.4ns ± 0%  12.2ns ± 1%   -8.84%  (p=0.016 n=4+5)
JumpdestOpAnalysis/PUSH27-6    15.1ns ± 0%  12.2ns ± 0%  -19.63%  (p=0.016 n=5+4)
JumpdestOpAnalysis/PUSH28-6    16.1ns ± 2%  12.1ns ± 4%  -24.74%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH29-6    17.5ns ± 6%  11.9ns ± 0%  -32.15%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH30-6    18.4ns ± 0%  11.6ns ± 5%  -36.93%  (p=0.016 n=4+5)
JumpdestOpAnalysis/PUSH31-6    19.7ns ± 1%  11.2ns ± 1%  -43.36%  (p=0.008 n=5+5)
JumpdestOpAnalysis/PUSH32-6    10.6ns ± 2%  11.5ns ± 0%   +8.23%  (p=0.016 n=5+4)
JumpdestOpAnalysis/JUMPDEST-6  38.9ns ± 5%  19.6ns ± 0%  -49.50%  (p=0.016 n=5+4)
JumpdestOpAnalysis/STOP-6      38.8ns ± 4%  19.6ns ± 0%  -49.51%  (p=0.016 n=5+4)

```